Move AccessStatus into its own file, separate from AccessCondition
